pretty: fix another bug in pretty printer
